Quick few changes
