+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/teleop/RobotTeleopMain.java
@@ -3,13 +3,9 @@ package org.firstinspires.ftc.teamcode.opmodes.teleop;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Servo;
 
-import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.teamcode.subsystems.ArmControl;
+import org.firstinspires.ftc.teamcode.drive.ArmSystem;
 import org.firstinspires.ftc.teamcode.subsystems.MechanumDrive;
-import org.firstinspires.ftc.teamcode.subsystems.TwoMotorTank;
-import org.firstinspires.ftc.vision.VisionPortal;
 
 @TeleOp(name = "TeleopMain")
 public class RobotTeleopMain extends OpMode {
@@ -19,14 +15,14 @@ public class RobotTeleopMain extends OpMode {
     DcMotor motorBackLeft;
 
 
-    //DcMotor motorLiftArm;
+    DcMotor motorLiftArm;
     //Servo servoGrabber;
     //Servo servoWrist;
     //Servo servoStopper;
 
 
     MechanumDrive driveSystem;
-    //ArmControl armSystem;
+    org.firstinspires.ftc.teamcode.drive.ArmSystem ArmSystem;
 
 
     //private VisionPortal visionPortal;
@@ -37,13 +33,13 @@ public class RobotTeleopMain extends OpMode {
         motorFrontRight = hardwareMap.dcMotor.get("motorFrontRight");
         motorBackLeft = hardwareMap.dcMotor.get("motorBackLeft");
         motorBackRight = hardwareMap.dcMotor.get("motorBackRight");
-        //motorLiftArm = hardwareMap.dcMotor.get("motorLiftArm");
+        motorLiftArm = hardwareMap.dcMotor.get("motorLiftArm");
         //servoGrabber = hardwareMap.servo.get("servoGrabber");
         //servoWrist = hardwareMap.servo.get("servoWrist");
         //servoStopper = hardwareMap.servo.get("servoStopper");
 
         driveSystem = new MechanumDrive(motorFrontLeft, motorFrontRight, motorBackLeft, motorBackRight);
-        //armSystem = new ArmControl(motorLiftArm, servoGrabber, servoWrist, servoStopper);
+        ArmSystem = new ArmSystem(motorLiftArm);
 
         //visionPortal = VisionPortal.easyCreateWithDefaults(
         //        hardwareMap.get(WebcamName.class, "Webcam 1"));
@@ -60,7 +56,7 @@ public class RobotTeleopMain extends OpMode {
 
         // First parameter: armUp input
         // Second parameter: armDown input
-        //armSystem.restrictedControlArmLift(gamepad2.right_bumper, gamepad2.right_trigger);
+        ArmSystem.restrictedControlArmLift(gamepad2.right_bumper, gamepad2.right_trigger);
 
         // First parameter: grabberOpen input
         // Second parameter: grabberClose input
